refactor(package-json): move utils functions to a separate file
